added withdrawing logic

+++ cmd/gophermart/controllers/Dto/Withdraw.go
@@ -1,6 +1,6 @@
 package Dto
 
 type Withdraw struct {
-	Order string `json:"order"`
-	Sum   int    `json:"sum"`
+	Order string  `json:"order"`
+	Sum   float64 `json:"sum"`
 }
